requirements.txt
python-telegram-bot==20.7
aiosqlite==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
nest_asyncio==1.6.0
loguru==0.7.2
requests==2.31.0

